% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fresh_project.R
\name{create_fresh_project}
\alias{create_fresh_project}
\title{Create folders for a new scientific data analysis project.}
\usage{
create_fresh_project(proj_name = "Fresh Project", proj_dir)
}
\arguments{
\item{proj_name}{What is the name of the project? This will be the name of the main folder for the project.}

\item{proj_dir}{A directory where you would like to set up the project.}
}
\value{
A new project named \code{proj_name} is created in the project directory \code{proj_dir}.

Examples Include:

create_fresh_project(proj_name = "Fresh Project", proj_dir = "/Users/USC/Documents/GitHub/freshproject")

create_fresh_project(proj_name = "My Fresh Project", proj_dir = "/Users/USC/Desktop")

create_fresh_project(proj_name = "Even Fresher Project", proj_dir = "/Users/USC/Desktop")

create_fresh_project(proj_name = "The Freshest Project", proj_dir = "/Users/USC/Desktop")
}
\description{
Create folders for a new scientific data analysis project.
}
